Module auraed::init

Run the Aurae daemon as a pid 1 init program.

The Aurae daemon assumes that if the current process id (PID) is 1 to run itself as an initialization program, otherwise bypass the init module.

Functions

init

Run Aurae as an init pid 1 instance.