Crate auraed

Aurae Daemon

Systems daemon built for higher order simple, safe, secure multi-tenant distributed systems.

Whether run as pid 1 (init), or a Container, or a Pod it serves standard library functionality over an mTLS backed gRPC server.

The Aurae Daemon (auraed) is the main server implementation of the Aurae Standard Library.

The Aurae Daemon runs as a gRPC server which listens over a unix domain socket by default.

/var/run/aurae/aurae.sock

Running Auraed

Running as /init is currently under active development.

To run auraed as a standard library server you can run the daemon alongside your current init system.

sudo -E auraed

See The Aurae Standard Library for API reference.

Structs

AuraedPath

Helper type to locate the auraed exe path.

AuraedRuntime

Each instance of Aurae holds internal state in memory. Below are the settings which can be configured for a given Aurae daemon instance.

Functions

prep_oci_spec_for_spawn

Write the container OCI spec to the filesystem in preparation for spawning Auraed using a container runtime.

run

Starts the runtime loop for the daemon.