Crate auraed

Systems daemon built for higher order simple, safe, secure multi-tenant distributed systems.

Runs as pid 1 (init) and serves standard library functionality over a mTLS backed gRPC server.

The Aurae Daemon (auraed) is the main server implementation of the Aurae Standard Library.

See The Aurae Standard Library for API reference.

Modules

init

Run the Aurae daemon as a pid 1 init program.

logging

Internal logging system for Auraed and all spawned Executables, Containers and Instances.

Functions

daemon